RavenDB-14548 Debug package analyzer - detection of long running transactions in System storage
